+++ user-application/src/main/java/io/github/dronesecurity/userapplication/reporting/negligence/entities/DroneData.java
@@ -26,7 +26,7 @@ public interface DroneData {
      * Gets the accelerometer data.
      * @return the accelerometer
      */
-    Map<String, Double> getAccelerometer();
+    Map<String, Integer> getAccelerometer();
 
     /**
      * Gets the camera data.
